Add open_path() Python method to permit persistent device symlinks.
Adds to the functionality provided by the open() call (which hard-codes
the opening of /dev/spidev<bus>.<chipselect> device file paths).
Because the bus number allocation may not be knowable in advance (bus
numbers are dynamically allocated by the Linux kernel), open_path()
allows the user to employ udev rules to deterministically create a
symlink to the correct hardware device at boot and/or hot-plug time.

Part of #129